RU: Методы CanUsePort() и GetAvailablePort() перенесены в новый класс NetHelper. Для тестов изменено пространство имен. Тестирование вышеперечисленных методов вынесено в отдельный файл NetHelperTests.
EN: The CanUsePort() and GetAvailablePort() methods have been moved to the new NetHelper class. The namespace for tests has been changed. Testing of the above methods is included in a separate file NetHelperTests.
